Delete srtc_util - turns out we have util.h/.cpp already
