+++ include/srtc/util.h
@@ -28,6 +28,8 @@ void getNtpTime(NtpTime& ntp);
 
 int64_t getStableTimeMicros();
 
+size_t compressNackList(const std::vector<uint16_t>& nackList, uint16_t* buf_seq, uint16_t* buf_blp);
+
 template <class T>
 class Filter
 {
