+++ src/util.cpp
@@ -141,6 +141,38 @@ int64_t getStableTimeMicros()
 #endif
 }
 
+size_t compressNackList(const std::vector<uint16_t>& nackList, uint16_t* buf_seq, uint16_t* buf_blp)
+{
+    const size_t in_end = nackList.size();
+
+    std::memset(buf_seq, 0, sizeof(uint16_t) * in_end);
+    std::memset(buf_blp, 0, sizeof(uint16_t) * in_end);
+
+    size_t out_pos = 0;
+    size_t in_pos = 0;
+
+    while (in_pos < in_end) {
+        const auto seq = nackList[in_pos];
+
+        buf_seq[out_pos] = seq;
+        buf_blp[out_pos] = 0;
+
+        in_pos += 1;
+
+        while (in_pos < in_end && static_cast<uint16_t>(nackList[in_pos] - seq) <= 16) {
+            const auto shift = static_cast<uint16_t>(nackList[in_pos] - seq - 1);
+            buf_blp[out_pos] |= 1u << shift;
+            in_pos += 1;
+        }
+
+        out_pos += 1;
+    }
+
+    return out_pos;
+}
+
+///// Filter
+
 template <class T>
 Filter<T>::Filter(float factor)
 	: mFactor(factor)
